requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==14.3.0
bitarray==3.3.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chess==1.11.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
cython==3.0.12
decorator==5.2.1
edge-tts==7.0.0
exceptiongroup==1.2.2
fairseq==0.12.2
faiss-cpu==1.7.3
fastapi==0.115.12
ffmpeg-python==0.2.0
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
h11==0.14.0
hydra-core==1.0.7
idna==3.10
importlib-resources==6.5.2
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.43.0
loguru==0.7.3
lxml==5.3.2
markupsafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
networkx==3.2.1
numba==0.60.0
numpy==1.23.5
omegaconf==2.0.6
openai==1.30.0
opencv-python==4.11.0.86
python-dotenv==1.0.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
portalocker==3.1.1
praat-parselmouth==0.4.5
propcache==0.3.1
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pyworld==0.3.5
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rvc-python==0.1.5
sacrebleu==2.5.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
srt==3.5.3
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
torch==2.6.0
torchaudio==2.6.0
torchcrepe==0.0.23
torchvision==0.21.0
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
ultralytics==8.3.107
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.0
yarl==1.19.0
zipp==3.21.0
